fix(output): fix macro OUT type error in output.h and fix paint Symbol error in output.c

+++ include/output.h
@@ -3,7 +3,7 @@
 
 #include "ast.h"
 
-#define OUT(indent, format, ...) fprintf(out, "%*s" format, (indent) * 4, "", ##__VA_ARGS__)
+#define OUT(indent, format, ...) fprintf(out, "%*s" format, ((int)indent) * 4, "", ##__VA_ARGS__)
 #define foreach(type, item, list) for (type item = (type)list_pop_front(list); item != NULL; item = (type)list_pop_front(list))
 
 void print_ast(AST* ast, FILE* out);
